# Improve performance of evaluate function