## Improving performance

Ok, this caught me by surprise. I noticed that  when I make the first move, this engine, which is not very good yet, was taking more than 10 seconds to evaluate only [3 levels deep](https://github.com/shermam/chess-engine/blob/bb47b1cd4f88bdad259c29648915e3b10e157380/worker/engine.js#L9). So I decided to grab a chrome dev tools performance panel snapshot to see what was going on. And I saw this:

![Screen Shot 2021-01-03 at 19 42 42](https://user-images.githubusercontent.com/12242359/103490655-0d968e00-4dfc-11eb-838b-9a6c7e4553e0.png)

You can see that the engine worker is taking 10.76 seconds to come up with a move. And looking a little bit down on the break down of the functions called, I saw that the [`evaluate` function](https://github.com/shermam/chess-engine/blob/9230ff2507a2a0b0e55d8168819a7ca324b5c03d/worker/engine.js#L14-L17) was responsible for 81.2% of that.

The evaluate function is not supposed to be expensive. I set up the board representation in such a way that the only thing I need to do to evaluate a position is sum an array of numbers. And here is how I was doing it:

```js
const evaluate = (position) => position.reduce((p, c) => p + c);
```

This was my first option because the one liner looked pretty to me. But somehow the multiliner good old `for` loop seems to improve the performance significantly:

```js
const evaluate = (position) => {
  let evaluation = 0;
  for (var i = 0; i < position.length; i++) {
    evaluation += position[i] ?? 0;
  }
  return evaluation;
};
```
Same exactly functionality dropped the task total time from 10.76s to 3.55s:

![Screen Shot 2021-01-03 at 19 52 02](https://user-images.githubusercontent.com/12242359/103490838-59960280-4dfd-11eb-928b-e68ca41dead2.png)

I did not look too deep into it yet to understand why such a difference. I is true that this function runs half a million times to calculate the best move. So even the smallest of improvements would have a huge impact on the outcome. So I guess that this improvement is the removal of the extra callback function. Maybe handling the function scope might be the thing introducing the significant overhead.

I know that there is a lot more to be improved here for this engine to be any good. But this simple change was really a surprise for me.